pyproject.toml
[project]
name = "aussiebb-blade-mcp"
version = "0.1.0"
description = "Aussie Broadband MCP Server — usage monitoring, outage tracking, and line diagnostics"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [{name = "Piers"}]
keywords = ["aussie-broadband", "nbn", "isp", "mcp", "broadband"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Networking :: Monitoring",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "fastmcp>=2.0.0",
    "pyaussiebb>=0.1.5",
    "httpx>=0.28.1",
    "pydantic>=2.0.0",
]

[dependency-groups]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
aussiebb-blade-mcp = "aussiebb_blade_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aussiebb_blade_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "e2e: end-to-end tests requiring live ABB account",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
show_error_codes = true
pretty = true
